Rename objects/attribute to objects/input
