+++ lib/inspec/objects/input.rb
