+++ lib/inspec/profile_context.rb
@@ -8,7 +8,7 @@ require 'inspec/library_eval_context'
 require 'inspec/control_eval_context'
 require 'inspec/require_loader'
 require 'securerandom'
-require 'inspec/objects/attribute'
+require 'inspec/objects/input'
 
 module Inspec
   class ProfileContext
